backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
chromadb
sentence-transformers
python-dotenv
cerebras-cloud-sdk
cryptography
